requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
python-multipart==0.0.6
httpx==0.25.2
asyncio==3.4.3
websockets==12.0
aiohttp==3.9.1
